+++ src/convmou.cpp
@@ -1,6 +1,17 @@
 #include "convmou.hpp"
+#include "dbimou.hpp"
 
 namespace mdbxmou {
+
+convmou convmou::for_dbi(const dbimou& dbi) noexcept
+{
+    return {
+        dbi.get_key_mode(),
+        dbi.get_value_mode(),
+        dbi.get_key_flag(),
+        dbi.get_value_flag()
+    };
+}
 
 Napi::Value convmou::convert_key(const Napi::Env& env, const keymou& key) const
 {
@@ -17,6 +28,12 @@ Napi::Value convmou::convert_key(const Napi::Env& env, const keymou& key) const
 
 Napi::Value convmou::convert_value(const Napi::Env& env, const valuemou& val) const
 {
+    if (is_ordinal(value_mode_)) {
+        return (value_flag_ & base_flag::bigint) ?
+            val.to_bigint(env) :
+            val.to_number(env);
+    }
+
     return (value_flag_ & base_flag::string) ?
         val.to_string(env) :
         val.to_buffer(env);