+++ src/convmou.hpp
@@ -4,20 +4,16 @@
 
 namespace mdbxmou {
 
+class dbimou;
+
 struct convmou
 {
     key_mode key_mode_{};
+    value_mode value_mode_{};
     base_flag key_flag_{};
     base_flag value_flag_{};
 
-    convmou() = default;
-
-    convmou(key_mode key_mode, base_flag key_flag,
-        base_flag value_flag = {}) noexcept
-        : key_mode_{key_mode}
-        , key_flag_{key_flag}
-        , value_flag_{value_flag}
-    {   }
+    static convmou for_dbi(const dbimou& dbi) noexcept;
 
     Napi::Value convert_key(const Napi::Env& env, const keymou& key) const;