+++ src/dbimou.cpp
@@ -335,7 +335,7 @@ Napi::Value dbimou::put(const Napi::CallbackInfo& info)
             keymou::from(info[1], env, t) : 
             keymou::from(info[1], env, key_buf_);
 
-        auto val = valuemou::from(info[2], env, val_buf_);
+        auto val = valuemou::from(info[2], env, val_buf_, val_num_, is_ordinal(value_mode_));
         MDBX_put_flags_t flags = MDBX_UPSERT;
         if (arg_len > 3 && !info[3].IsUndefined() && !info[3].IsNull()) {
             if (!info[3].IsNumber()) {